requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
backoff==2.2.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cfgv==3.5.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
decouple==0.0.7
distlib==0.4.0
Django==5.2.11
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-crontab==0.7.1
django-environ==0.12.0
django-filter==25.2
django-ipware==7.0.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
Faker==28.4.1
filelock==3.20.3
frozenlist==1.8.0
geoip2==5.2.0
gql==4.0.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==25.0.2
identify==2.6.16
idna==3.11
inflection==0.5.1
kombu==5.6.2
Markdown==3.10
maxminddb==3.0.0
multidict==6.7.0
mysqlclient==2.2.7
nodeenv==1.10.0
packaging==25.0
pillow==12.1.0
platformdirs==4.5.1
pre_commit==4.5.1
promise==2.3
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
reportlab==4.2.5
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.10.1
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
virtualenv==20.36.1
wcwidth==0.2.14
wheel==0.46.2
whitenoise==6.11.0
yarl==1.22.0
